Deprecate `bcn run` command
