+++ cmd/deploy_test.go
@@ -1,9 +1,11 @@
 package cmd
 
 import (
+	"io/ioutil"
 	"os"
 
 	"github.com/jarcoal/httpmock"
+	"github.com/urfave/cli"
 )
 
 func Example_deploy_quiet_false() {
@@ -56,4 +58,28 @@ func Example_deploy_quiet_true() {
 	app.Run(testArgs)
 
 	// Output:
+}
+
+func newTestApp(command cli.Command) *cli.App {
+	a := cli.NewApp()
+	a.Name = "bcn"
+	a.Writer = ioutil.Discard
+	a.Commands = []cli.Command{command}
+
+	pwd, _ := os.Getwd()
+	HeritageConfigFilePath = pwd + "/test/test-barcelona.yml"
+
+	return a
+}
+
+func readJsonResponse(path string) (string, error) {
+	jsonFile, err := os.Open(path)
+	if err != nil {
+		return "", err
+	}
+
+	byteValue, _ := ioutil.ReadAll(jsonFile)
+	defer jsonFile.Close()
+
+	return string(byteValue), err
 }
